back_end/requirements.txt
 # FastAPI Framework
fastapi
uvicorn[standard]

# Database
sqlalchemy
psycopg2-binary
alembic
asyncpg

# File Handling
python-multipart
#pandas==2.1.3
openpyxl

# AI/ML


# Environment & Security
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
python-multipart

# CORS
#fastapi-cors==0.1.0

# Utilities
pydantic-settings
pydantic
redis
celery

# Development
pytest
pytest-asyncio
black
flake8
